feat:topKeyword 5개 적용

+++ back/src/main/java/com/cocoa/backend/domain/cosmetic/entity/CosmeticKeywords.java
@@ -15,7 +15,7 @@ import lombok.Getter;
 @Table(name = "cosmetic_keywords")
 public class CosmeticKeywords {
     @Id
-    private Long cosmeticId;
+    private Integer cosmeticId;
 
     @Column(columnDefinition = "jsonb")
     @Convert(converter = KeywordJsonConverter.class)
